requirements.txt
Flask==3.0.3
pymongo==4.8.0
pydantic==2.8.2
python-dotenv==1.0.1
Flask-Cors==4.0.1
email-validator==2.2.0
